Update LICENSE.md
Take My Photo Down
(http://takemyphotodown.herokuapp.com/)

	Take My Photo Down is a website that allows victims of revenge pornography to use the Digital Millennium Copyright Act (DMCA) to remove “selfie” photographs from specific websites. Those who take photographs of themselves obtain a copyright interest in the photograph and Take My Photo Down provides people a means to use this copyright interest to get their photos taken off websites which may be hosting them. The site allows people to submit information about themselves, the photograph, and the website, and then generates an email which alerts the website that it is currently hosting copyright infringing content. Because of the problems that revenge porn has caused in recent years, states, prosecutors, and private attorneys have attempted to address the problem in different ways, such as criminalizing the posting of revenge porn or bringing suits for harassment or intentional infliction of emotional distress. By using the DMCA and providing an easy-to-use take-down notification generator, Take My Photo Down offers an alternative which may be more efficient while at the same time being less costly and less intrusive. 

	The DMCA shields internet service providers (ISPs) from liability when users upload copyrighted material onto the sites and services they operate. This statute helps websites like YouTube to operate without liability despite the fact that users may upload copyright infringing content. The ISPs only receive the benefit of this safe harbor, however, if they abide by the statute’s take-down provision. To be DMCA compliant, an ISP must first register a DMCA agent with the Copyright Office. This agent then must remove copyrighted material from the site when he or she is given adequate notice that the material is being hosted. Because those who take “selfie” photographs obtain a copyright interest in the photos, they are able to use the DMCA to get those photos taken down, and, because the safe harbor is essential for content-hosting ISPs to operate, these ISPs are very incentivized to take take-down notifications seriously. 

The Take My Photo Down project was begun by former BLIP students and has been revived this semester. The website already exists in a very basic form—users can input their information and the site will generate a DMCA takedown notice to send to an offending website. Some minor improvements to the Take My Photo Down framework, however, may vastly improve the website. Although the site does contain some background information about revenge porn and the DMCA, the information is minimal and somewhat uninstructive, especially for those with no understanding of copyright law. We would like to add content to the site which greater explains the revenge porn problem, the options currently available to address it, the DMCA take-down provision, and the benefits and detriments of using the DMCA as a remedy. Users need to be aware of what exactly Take My Photo Down does, how it can be helpful, and how it should not be used. By adding additional content and links to helpful information elsewhere on the internet, we can make the site more accessible and the process less intimidating. 

The Take My Photo Down framework is well suited to helping victims of revenge porn remove photos of themselves from the internet—the process is simple and should not require any technical or legal knowledge to utilize. The framework, however, can be expanded in several ways. First, although the project aimed at providing a solution for the revenge porn problem, the take-down notification generator can be used in any case of copyright infringement, and may be helpful for artists, photographers, and musicians who do not have the resources to hire legal counsel to get infringing content removed from websites. The concept of a website generating a form email can also be helpful outside the copyright context. Similar websites can be developed to generate FOIA requests or expungement requests. The goal here was to provide a workable solution to a common and growing problem, but there is no reason that the solution created here should be limited specifically to revenge pornography when there are many ways that the project can be expanded.